backend/requirements.txt
asgiref==3.2.3
Django==3.0
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
requests==2.27.1
django-cors-headers==3.11.0